requirements.txt
numpy==1.24.4
pandas==2.0.3
PyYAML==6.0
moderngl==5.7.4
PyOpenGL==3.1.7
glfw==2.5.5
imgui==2.0.0
PyGLM==2.7.1
Pillow==10.0.0